refactor(backend-game-adapter-typeorm): move GameInitialSnapshotDb models
