+++ packages/backend/apps/game/backend-game-adapter-typeorm/src/foundation/db/adapter/nest/models/entities.ts
@@ -1,7 +1,7 @@
+import { GameInitialSnapshotDb } from '../../../../../gameInitialSnapshots/adapter/typeorm/models/GameInitialSnapshotDb';
+import { GameInitialSnapshotSlotDb } from '../../../../../gameInitialSnapshots/adapter/typeorm/models/GameInitialSnapshotSlotDb';
 import { GameActionDb } from '../../../../../games/adapter/typeorm/models/GameActionDb';
 import { GameDb } from '../../../../../games/adapter/typeorm/models/GameDb';
-import { GameInitialSnapshotDb } from '../../../../../games/adapter/typeorm/models/GameInitialSnapshotDb';
-import { GameInitialSnapshotSlotDb } from '../../../../../games/adapter/typeorm/models/GameInitialSnapshotSlotDb';
 import { GameSlotDb } from '../../../../../games/adapter/typeorm/models/GameSlotDb';
 import { GameSpecDb } from '../../../../../games/adapter/typeorm/models/GameSpecDb';
