+++ packages/backend/apps/game/backend-game-adapter-typeorm/src/gameInitialSnapshots/adapter/typeorm/models/GameInitialSnapshotSlotDb.ts
